add test into a seperate file
